requirements.txt
# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.15.0

# Google Sheets Integration
gspread>=5.10.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Date/Time Handling
python-dateutil>=2.8.0

# HTTP Requests (fallback for CSV)
requests>=2.31.0

# Development Tools (optional)
python-dotenv>=1.0.0